+++ src/Symfony/Component/Routing/Tests/Fixtures/validpattern.php
@@ -6,7 +6,7 @@ use Symfony\Component\Routing\RouteCollection;
 $collection = new RouteCollection();
 $collection->add('blog_show', new Route(
     '/blog/{slug}',
-    ['_controller' => 'MyBlogBundle:Blog:show'],
+    ['_controller' => 'MyBlogBundle:Blog:show', '_stateless' => true],
     ['locale' => '\w+'],
     ['compiler_class' => 'RouteCompiler'],
     '{locale}.example.com',