+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
  * added argument `$priority` to `RouteCollection::add()`
  * deprecated the `RouteCompiler::REGEX_DELIMITER` constant
  * added `ExpressionLanguageProvider` to expose extra functions to route conditions
+ * added support for a `stateless` keyword for configuring route stateless in PHP, YAML and XML configurations.
 
 5.0.0
 -----